requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3
cycler==0.10.0
Cython==0.25.2
diff-match-patch==20121119
Django==1.10.5
django-bootstrap3==8.0.0
django-bootstrap3-datetimepicker==2.2.3
django-bootstrap3-datetimepicker-2==2.4.2
django-bootstrap4==0.0.4
django-import-export==1.0.1
django-mysql==2.4.0
django-smtp-ssl==1.0
et-xmlfile==1.0.1
future==0.15.2
idna==2.7
jdcal==1.4
Jinja2==2.8
MarkupSafe==0.23
matplotlib==1.5.3
mysql-connector-python==2.0.4
numpy==1.11.0
odfpy==1.3.6
opencv-python==3.2.0.6
openpyxl==2.5.5
pefile==2016.3.28
pycparser==2.18
PyMySQL==0.9.2
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.13
requests==2.10.0
six==1.10.0
suds-jurko==0.6
tablib==0.12.1
unicodecsv==0.14.1
virtualenv==15.0.1
Werkzeug==0.11.11
